requirements.txt
nltk~=3.9.1
pymorphy3~=2.0.4
bitarray~=3.4.2
PyYAML~=6.0.1
pandas~=2.2.3
Pympler~=1.1